Refactor UI: extract style palette dialog module
